pyproject.toml
[project]
name = "voc-project"
version = "0.1.0"
description = "VOC monitoring application with PySide6 GUI"
requires-python = ">=3.11"
dependencies = [
    "PySide6>=6.6.0",
    "numpy>=1.24.0",
    "pyserial>=3.5",
]

[project.optional-dependencies]
rpi = [
    "RPi.GPIO>=0.7.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
security = [
    "bcrypt>=4.0.0",
    "argon2-cffi>=23.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/voc_app"]
